fix: set 'cluster-init: false' when explicitly disabled via ProviderOptions
